Publication in Energy Procedia
This publication in Elsevier's Energy Procedia replaces the draft publication of the paper submitted for the IRES 2018. The new publication is longer and more complete. It contains notably sensitivity analysis in more detail.
For wider dissemination of this publication, please, use the link to Elsevier: https://www.sciencedirect.com/science/article/pii/S1876610218310117?via%3Dihub